+++ shardingsphere-infra/shardingsphere-infra-common/src/main/java/org/apache/shardingsphere/infra/metadata/schema/builder/physical/ColumnMetaDataLoader.java
@@ -36,10 +36,10 @@ import java.util.List;
 import java.util.Objects;
 
 /**
- * Physical column meta data loader.
+ * Column meta data loader.
  */
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
-public final class PhysicalColumnMetaDataLoader {
+public final class ColumnMetaDataLoader {
     
     private static final String COLUMN_NAME = "COLUMN_NAME";
