+++ shardingsphere-infra/shardingsphere-infra-common/src/test/java/org/apache/shardingsphere/infra/metadata/schema/builder/physical/IndexMetaDataLoaderTest.java
@@ -17,46 +17,45 @@
 
 package org.apache.shardingsphere.infra.metadata.schema.builder.physical;
 
-import org.apache.shardingsphere.infra.database.type.DatabaseType;
-import org.junit.Before;
+import org.apache.shardingsphere.infra.metadata.schema.model.IndexMetaData;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.Answers;
-import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnitRunner;
 
-import javax.sql.DataSource;
+import java.sql.Connection;
 import java.sql.ResultSet;
 import java.sql.SQLException;
-import java.util.Collections;
+import java.util.Collection;
 
 import static org.hamcrest.CoreMatchers.is;
 import static org.junit.Assert.assertThat;
+import static org.mockito.Mockito.RETURNS_DEEP_STUBS;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
 @RunWith(MockitoJUnitRunner.class)
-public final class PhysicalSchemaMetaDataLoaderTest {
+public final class IndexMetaDataLoaderTest {
     
-    @Mock(answer = Answers.RETURNS_DEEP_STUBS)
-    private DataSource dataSource;
+    @Test
+    public void assertLoad() throws SQLException {
+        Collection<IndexMetaData> actual = IndexMetaDataLoader.load(mockConnection(), "tbl");
+        assertThat(actual.size(), is(1));
+        IndexMetaData indexMetaData = actual.iterator().next();
+        assertThat(indexMetaData.getName(), is("my_index"));
+    }
     
-    @Before
-    public void setUp() throws SQLException {
+    private Connection mockConnection() throws SQLException {
+        Connection result = mock(Connection.class, RETURNS_DEEP_STUBS);
         ResultSet resultSet = mockResultSet();
-        when(dataSource.getConnection().getMetaData().getTables("catalog", null, null, new String[]{"TABLE", "VIEW"})).thenReturn(resultSet);
-        when(dataSource.getConnection().getCatalog()).thenReturn("catalog");
+        when(result.getMetaData().getIndexInfo("catalog", null, "tbl", false, false)).thenReturn(resultSet);
+        when(result.getCatalog()).thenReturn("catalog");
+        return result;
     }
     
     private ResultSet mockResultSet() throws SQLException {
         ResultSet result = mock(ResultSet.class);
-        when(result.next()).thenReturn(true, true, true, false);
-        when(result.getString("TABLE_NAME")).thenReturn("tbl", "$tbl", "/tbl");
+        when(result.next()).thenReturn(true, true, false);
+        when(result.getString("INDEX_NAME")).thenReturn("my_index");
         return result;
-    }
-    
-    @Test
-    public void assertLoadAllTableNames() throws SQLException {
-        assertThat(PhysicalSchemaMetaDataLoader.loadAllTableNames(dataSource, mock(DatabaseType.class)), is(Collections.singletonList("tbl")));
     }
 }
